pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "A powerful terminal-native AI assistant with tool execution"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "prompt-toolkit>=3.0.0",
    "requests>=2.0.0",
    "python-dotenv>=1.0.0",
    "pygments>=2.0.0",
    "uvloop>=0.22.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.6",
]

[project.scripts]
agent = "src.agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 120
